@@ -1,30 +1,28 @@
-// https://eslint.org/docs/user-guide/configuring
-
 module.exports = {
   root: true,
-  parserOptions: {
-    parser: "babel-eslint"
-  },
   env: {
-    node: true,
-    browser: true
+    node: true
   },
-  extends: [
-    // https://github.com/vuejs/eslint-plugin-vue#priority-a-essential-error-prevention
-    // consider switching to `plugin:vue/strongly-recommended` or `plugin:vue/recommended` for stricter rules.
-    "plugin:vue/essential",
-    // https://github.com/standard/standard/blob/master/docs/RULES-en.md
-    "standard",
-    "eslint:recommended"
+  'extends': [
+    'plugin:vue/vue3-essential',
+    'eslint:recommended'
   ],
-  // required to lint *.vue files
-  plugins: ["vue"],
-  // add your custom rules here
+  parserOptions: {
+    parser: 'babel-eslint'
+  },
   rules: {
-    // allow async-await
-    "generator-star-spacing": "off",
-    // allow debugger during development
-    "no-console": process.env.NODE_ENV === "production" ? "error" : "off",
-    "no-debugger": process.env.NODE_ENV === "production" ? "error" : "off"
-  }
-};
+    'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
+    'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off'
+  },
+  overrides: [
+    {
+      files: [
+        '**/__tests__/*.{j,t}s?(x)',
+        '**/tests/unit/**/*.spec.{j,t}s?(x)'
+      ],
+      env: {
+        jest: true
+      }
+    }
+  ]
+}
